streamlit/app/requirements.txt
streamlit>=1.28.0
streamlit-extras>=0.3.0
streamlit-authenticator>=0.2.1
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.31.0
PyYAML>=6.0
python-dotenv>=1.0.0
mlflow>=2.5.0
scikit-learn>=1.3.0
joblib>=1.2.0
psycopg2-binary>=2.9.6
SQLAlchemy>=2.0.0
python-dateutil>=2.8.2
pytz>=2023.3
typing-extensions>=4.5.0
python-multipart>=0.0.5
